Fix nodeBuds paragraph alignment, turnback class
This commit fixes the second paragraph's alignment in mobile view to
render properly. It also fixes a minor capitalization error ("nodebuds"
to "nodeBuds").

This commit also adds the groundwork for potential future refactoring:
the CSS variables --fluid-font-size, --p-font-size and --h-font-size-med
were introduced to slowly replace the previous fluidSize mixin.

The --fluid-font-size uses the CSS clamp() math function to achieve
smoothly scalable font sizes across all viewports unlike setting manual
font-sizes over @media. This behavior is almost similar to the mixin.
Its intended use case is to be set at the parent container (or the
body). Note that this affects the margins and paddings if they use em as
well.

The two new font size variables, --p-font-size and --h-font-size-med,
are in em units to properly utilize the above --fluid-font-size set in
the parent container. The em units will scale accordingly. This is
different from using px, which is scaled absolutely (as opposed to
relatively to the parent's font-size).