Meeting with Designer
The changes in this commit are the result of the notes of a meeting with the site's designer.

## Structural Changes

For component sections of the index page that are not used anywhere besides the index page have been moved to a new directory, `/src/components/index`. This includes `acm-paths.svelte`, `hero.svelte`, and `why-join.svelte`.

For the NodeBuds page, some unique section components were also moved into a new directory, `/src/components/nodebuds`. This includes `nodebuds-testimonial.svelte` and `why-join-nodebuds.svelte`. Take note that  I also changed the URL of the NodeBuds endpoint from `node-buds` to `nodebuds` since it is written as one word.

Additionally, the `src/components/ui` directory has been renamed to `src/components/utils`.

## Stylistic Changes

New variables have been added to the `global.css` file!

The following variables have to do with the `font-size` style property and are intended to be used everywhere throughout the site.

```css
:root {
  --heading-font-size: 48px;
  --subheading-font-size: 24px;
  --body-font-size: 18px;
  --margin-text-font-size: 12px;
}

@media screen and (min-width: 768px) {
  :root {
    --title-font-size: 96px;
    --heading-font-size: 64px;
    --subheading-font-size: 36px;
    --body-font-size: 24px;
    --margin-text-font-size: 16px;
  }
}
```

The values that the variables are set to outside of the media query are the values used in mobile mode. The values set inside of the media query are the values used in tablet/desktop mode.

In a number of files that are used on the index page, any unnecessary padding has been eliminated in favor of the `Spacing` component. This change contributes to issue #56, but does not complete it.

Also, all of the instances on the index page where some sort of max-width is used (as a constraint to help with the page's layout) were replaced with margins instead.