Added an isLoading message
Added an isLoading feature so that the user is not confused by the "There are currently no events scheduled." message while they wait for the data to load. I wanted to use this [await block](https://svelte.dev/docs#await), but it felt right to just leave the loading in the `onMount` callback.