Added replaceHtmlWithExternalLinks function to utils.ts (#431)
* Update utils.ts

Added the assertHtmlHasAllExternalLinks function to utils.ts and it is called inside the parseDescription function. Testing still needs to be done on the assertHtmlHasAllExternalLinks function.

* Update utils.ts

Changed function name from assertHtmlHasAllExternalLinks to replaceHtmlWithExternalLinks.

* Made changes to utils.test.ts and utils.ts

Added a test to utils.test.ts and made progress on replaceHtmlWithExternalLinks. Still having issues with string manipulation.

* Updated utils.ts and utils.test.ts

Incorporated @jaasonw regex code and testcases.

* Updated utils.test.ts and utils.ts

Fixed up the regex code to exclude any tag that is not the anchor tag.

* Update utils.test.ts

New testcase added