Playwright testing: Remove screenshot testing (#637)
* Remove screenshot testing (besides 404 page)

The acmcsuf.com full page screenshot testing experiment has come to an end (for now). In its current state, the screenshot testing is like my hair: a bit long and cool, but always getting in the way! For this reason, I will delete the screenshot tests from the `main` branch but I will leave the individual Playwright page tests for future testing.

* Run `npm i && npm run all`