Implement officer legacyPicture workaround
This commit implements a workaround to fit board members with legacy
pictures onto the new <BoardPicture> component. It does this with
hard-coded translations to overlap the new frame onto the old pictures.

This hard-coding is written with the expectation that no legacy pictures
will be added in the future. It's not guaranteed to work with new legacy
pictures.