Add support for ics.acmcsuf.com
This commit adds support for parsing ICS calendars using the ICal
standard format in addition to the existing Google Calendar API.

The new commit also contains a hard-coded array of ICS calendar links,
currently only containing ics.acmcsuf.com. In the future, more links may
be added to this.

A hard-coded array was chosen over an environment variable because
environment variables cannot express arrays as easily. That said, if
there are any private calendars, then they can be used as such:

    const KnownCalendars = [
      "public1.ics",
      "public2.ics",
      SECRET_ICS_X,
      SECRET_ICS_Y,
    ]

In the future, if ics.acmcsuf.com is proven to work well enough, the
Google Calendar API could even be completely phased out.

NOTE: It is worth pointing out that this ICS code currently *does not
support recurring events*. This is because ics.acmcsuf.com doesn't do
recurring events.